Fixes #29
This is one of those magic moments where you can’t believe computers
behave this way. Surely, someone shouldn’t have to write JavaScript
like this, but somehow it works to force synchonous loading in an
asynchronous environment.

Committing this to GitHub because it seems like magic, but at the same
time, does seem to work both in Chrome and Safari.